# Improve speed of random_agent