--- a/docs/src/api.md
+++ b/docs/src/api.md
@@ -13,6 +13,7 @@ UnremovableABM
 ```@docs
 getindex(::ABM, ::Integer)
 getproperty(::ABM, ::Symbol)
+random_id
 random_agent
 nagents
 allagents