@Tortar 2023-10-05
I have to say that this means that probably there is something to improve in Base, since that the previous method is so slow sounds strange