@Tortar 2023-10-05
Actually it was a O(N) operation while this one is not, this means that the improvement in speed is arbitrary large!